Moved chess logic under common for server and client
